api/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.3
djangorestframework==3.12.4
idna==3.3
Jinja2==3.0.2
kiwisolver==1.3.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
mpld3==0.5.5
numpy==1.21.2
gunicorn==19.7.1
pandas==1.3.3
Pillow==8.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-image==0.18.3
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
unidecode==1.3.2
